♻️ refactor: 주문 상세 내역 api 수정
